project/requirements.txt
Django~=5.0.7
djangorestframework~=3.15.2
psycopg2~=2.9.9
django-environ
gunicorn
requests
PyJWT